Refactor Admin Dashboard Transaction, created Transaction Table component
